fix(runtime): address sourcery follow-ups
