+++ docs/review/PR_1010_FIXED_MAPPING.md
@@ -5,4 +5,22 @@
 - [x] Fixed in commit mapping completed
 
 ## Fixed in Commit Mapping
-- No actionable review comments
+Disposition: NOT-A-BUG
+Evidence: AGENTS.md:349
+Reason: Docs Phase1 gates explicitly require `file:line` evidence anchors, so replacing them with only section anchors would violate the current canonical docs contract.
+- https://github.com/Katsiarynakavaleuskaya/PulsePlate/pull/1010#pullrequestreview-3909426981
+
+Disposition: FIXED
+Commit: <pending>
+Evidence: app/models/llm_quota_usage.py:17
+- https://github.com/Katsiarynakavaleuskaya/PulsePlate/pull/1010#discussion_r2900340993 -> <pending>
+
+Disposition: FIXED
+Commit: <pending>
+Evidence: docker-compose.yaml:19
+- https://github.com/Katsiarynakavaleuskaya/PulsePlate/pull/1010#discussion_r2900340994 -> <pending>
+
+Disposition: FIXED
+Commit: <pending>
+Evidence: tests/test_rag_contract_surface.py:1
+- https://github.com/Katsiarynakavaleuskaya/PulsePlate/pull/1010#discussion_r2900340996 -> <pending>
